+++ ruzstd/src/decoding/mod.rs
@@ -10,6 +10,7 @@ pub use streaming_decoder::StreamingDecoder;
 
 pub(crate) mod block_decoder;
 pub(crate) mod decode_buffer;
+pub(crate) mod flat_buffer;
 pub(crate) mod dictionary;
 pub(crate) mod frame;
 pub(crate) mod literals_section_decoder;